requirements.txt
Flask
pandas
gunicorn
pytz
werkzeug
Flask-Login
psycopg2-binary # ถ้าใช้ PostgreSQL
python-dotenv # เพิ่มบรรทัดนี้
cloudinary # เพิ่มบรรทัดนี้
openpyxl
xlsxwriter
Flask-Caching
redis
python-barcode
qrcode[pil]
reportlab
requests
Pillow
sentry-sdk
gspread
oauth2client
Flask-Cors